# dist/api.yml
...
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:3000/
    description: SwaggerHub API Auto Mocking
...
paths:
...
  /game/allGames/:
    get:
      tags:
        - Games
      summary: return all the games in the season 
      responses:
        "200":
          description: match
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
        "400":
          description: something went wrong
  /game/getClosestGame:
    get:
      tags:
        - Games
      summary: returns the colset game in the leauge
      responses:
        "200":
          description: match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "400":
          description: something went wrong
  /game/getCurrentStageGames:
    get:
      tags:
        - Games
      summary: returns the current stage games
      responses:
        "200":
          description: match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "400":
          description: something went wrong
  /game/getDetails/{gameId}:
    get:
      tags:
        - Games
      summary: search a game by ID
      parameters:
        - name: gameId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: 42
      responses:
        "200":
          description: match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "400":
          description: something went wrong
  /gamechange/: #changed
    post:
      tags:
      - Games, Admins
      summary: Add a new game to the server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gameDate:
                  type: date
                  pattern: /([0-9]{4})-(?:[0-9]{2})-([0-9]{2})/
                  example: "2021-05-30"
                homeTeam:
                  type: integer
                  example: 85
                awayTeam:
                  type: 390
                  example: 123
                referee:
                  type: String
                  example: John
                stageID:
                  type: integer
                  example: 6
                stadium: 
                  type: String
                  example: Silverdo
      responses:
        "201":
          description: Game created
        "409":
          description: Game already in system
        "400":
          description: something went wrong
  /gamechange/{gameId}:
    put: #change
      tags:
      - Games, Admins
      summary: update game results
      parameters:
      - name: gameId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: 56
      requestBody:
        description: game results
        content:
          application/json:
            schema:
              type: object
              properties:
                homeScore:
                  type: integer
                awayScore:
                  type: integer
        required: true
      responses:
        "200":
          description: game updated
        "400":
          description: something went wrong
  /game/{gameId}/events/:
    get: #CHANGED
      tags:
      - Games
      summary: return all the event of the game
      parameters:
      - name: gameId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: 56 
      responses:
        "200":
          description: OK. recipes on the way.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'

        "204":
          description: Bad, No Event was found.
        "406":
          description: Game Not Found.
        "401":
          description: Unauthorized.
        "400":
          description: Bad Request.
  /gamechange/{gameId}/events/:
    post:
      tags:
      - Games, Admins
      summary: Add an event to a game
      parameters:
      - name: gameId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: 56
      requestBody:
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Event'
        required: true
      responses:
        "201":
          description: event added sucssfuly.
        "406":
          description: Game Not Found.
        "400":
          description: Bad Request.
        "401":
          description: Unauthorized.
  /coach/coachDetailById/{coachId}:
    get:
      tags:
        - Coachs
      summary: get coach personal page by ID
      operationId: CoachlPlayerPage
      parameters:
        - name: coachId
          in: path
          description: Identification of the requested coach by id.
          required: true
          style: simple
          explode: false
          schema:
            type: number
            example: 456421
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coach"
        "204":
          description: Bad, No Content was found.
        "400":
          description: Bad Request.
        "401":
          description: Unauthorized.

    # put:
    #   tags:
    #   - Coachs
    #   summary: get coach personal page
    #   operationId: CoadIDchange
    #   parameters:
    #   - name: coachId
    #     in: path
    #     description: Identification of the requested coach by id.
    #     required: true
    #     style: simple
    #     explode: false
    #     schema:
    #       type: number
    #       example: 234
    #   responses:
    #     "200":
    #       description: success
    #     "400":
    #       description: somting went wrong
  /coach/coachDetailByName/{coach_name}:
    get:
      tags:
        - Coachs
      summary: search coach by name by name
      parameters:
        - name: coach_name
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: Jens
      responses:
        "200":
          description: data of specific user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coach"

        "400":
          description: Bad Request.
  /coach/coachDetailByName/{coach_name}/filterbyTeam/{teamName}:
    get:
      tags:
        - Coachs
      summary: Search player by name
      parameters:
        - name: coach_name
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: J
        - name: teamName
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
          example: Horsens
      responses:
        "200":
          description: data of specific player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "400":
          description: Bad Request.
...
components:
  schemas:
...
    Referee:
      type: object
      properties:
        id:
          type: integer
          example: 33
        Name:
          type: string
          example: Maccabi Tel Aviv
        image-url:
          type: string
          format: URL
        qualification:
          type: string
          example: Line max
...
    Player:
      type: object
      properties:
        id:
          type: integer
          example: 82808
        fullName:
          type: string
          example: Jesper Hansen
        teamName:
          example: Midtjylland
        image-url:
          type: string
          format: URL
          example: https://cdn.sportmonks.com/images/soccer/players/24/82808.png
        position:
          type: integer
          example: 1
    Coach:
      type: object
      properties:
        id:
          type: integer
          example: 456421
        fullName:
          type: string
          example: Jess Thorup
        teamName:
          type: string
          example: København
        image-url:
          type: string
          format: URL
    Team:
      type: object
      properties:
        id:
          type: integer
          example: 85
        teamName:
          type: string
          example: København
        coach-id:
          type: integer
          example: 456421
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        Games:
          type: array
          items:
            $ref: "#/components/schemas/Game"
    Game:
      type: object
      properties:
        gameDate:
          type: string
          example: 19941030
        homeTeam:
          $ref: "#/components/schemas/Team"
        awayTeam:
          $ref: "#/components/schemas/Team"
        referees:
          $ref: "#/components/schemas/Referee"
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        homeScore:
          type: integer
          example: 2
        awayScore:
          example: 3
    Event:
      type: object
      properties:
        gameId:
          type: integer
          example: 12
        eventType:
          type: string
          example: "red Card"
        gameDate:
          type: string
          example: 19941030
        gameTime:
          type: string
          example: "17:30:25"
        inGameMinute:
          type: integer
          example: 13
        eventDescription:
          type: string
          example: goal by Ronaldo
...